Refactor contexts
