+++ src/components/Header.tsx
@@ -1,7 +1,7 @@
 import classNames from 'classnames'
 import React, { HTMLAttributes } from 'react'
 import { Link } from 'react-router-dom'
-import { useLocalizedPath, useLocalizedString } from '../contexts/i18n'
+import { useLocalizedPath, useLocalizedString } from '../providers/i18n'
 import style from './Header.module.css'
 
 type Props = HTMLAttributes<HTMLElement>
