+++ src/providers/i18n/hooks/useResetLocale.ts
@@ -1,7 +1,7 @@
 import { useContext } from 'react'
 import { useLocation, useNavigate } from 'react-router'
-import { I18nContext } from '../providers/I18nProvider'
-import { I18nRouterContext } from '../providers/I18nRouterProvider'
+import { I18nContext } from '../I18nProvider'
+import { I18nRouterContext } from '../I18nRouterProvider'
 import { getUnlocalizedURL } from '../utils/urls'
 
 /**
