+++ src/providers/i18n/index.ts
@@ -1,2 +1,3 @@
+export * from './hooks'
 export { default as I18nProvider } from './I18nProvider'
 export { default as I18nRouterProvider } from './I18nRouterProvider'
